frontend/backend/requirements.txt
# Workspace version: 1.49.4
# databutton==0.38.34  # Removed - migrating to Google Cloud Secret Manager and Firebase Storage
ipykernel==6.29.5
fastapi==0.115.12
python-multipart==0.0.9
uvicorn[standard]==0.34.2
gunicorn>=21.0.0,<24

numpy>=1.24.0,<3
audioop-lts>=0.2.1
openai>=1.0.0,<2
beautifulsoup4>=4.12.0
requests>=2.28.0,<3
google-genai>=1.0.0
pydub>=0.25.0
firebase-admin>=6.0.0
Pillow>=10.0.0
python-docx>=1.0.0
reportlab>=4.0.0
httpx>=0.25.0
pydantic[email]>=2.0.0
PyJWT>=2.8.0
google-cloud-translate>=3.0.0
google-cloud-speech>=2.0.0
resend>=0.8.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-cloud-secret-manager>=2.0.0
google-cloud-tasks>=2.0.0
jiwer==3.0.4
rapidfuzz==3.9.7
